👾 Add gltf model and apply Decal drei
🙂 ‍I have not yet figured out how to do this with only shaders.
👍 I can apply decal on plane geometry with normal + displacement map
to show the shapes, but not sure how to apply on gltf geometry.
📝 Maybe need to pass gltf geometry to shader as uniforms.
