+++ public/models/lemon/lemon_1k.gltf
@@ -1,0 +1,163 @@
+{
+    "asset": {
+        "generator": "Khronos glTF Blender I/O v4.1.63",
+        "version": "2.0"
+    },
+    "extensionsUsed": ["KHR_materials_specular", "KHR_materials_ior"],
+    "scene": 0,
+    "scenes": [
+        {
+            "name": "Scene",
+            "nodes": [0]
+        }
+    ],
+    "nodes": [
+        {
+            "mesh": 0,
+            "name": "lemon"
+        }
+    ],
+    "materials": [
+        {
+            "doubleSided": true,
+            "extensions": {
+                "KHR_materials_specular": {
+                    "specularColorFactor": [2, 2, 2]
+                },
+                "KHR_materials_ior": {
+                    "ior": 1.4500000476837158
+                }
+            },
+            "name": "lemon",
+            "normalTexture": {
+                "index": 0
+            },
+            "pbrMetallicRoughness": {
+                "baseColorTexture": {
+                    "index": 1
+                },
+                "metallicFactor": 0,
+                "metallicRoughnessTexture": {
+                    "index": 2
+                }
+            }
+        }
+    ],
+    "meshes": [
+        {
+            "name": "lemon_01",
+            "primitives": [
+                {
+                    "attributes": {
+                        "POSITION": 0,
+                        "NORMAL": 1,
+                        "TEXCOORD_0": 2
+                    },
+                    "indices": 3,
+                    "material": 0
+                }
+            ]
+        }
+    ],
+    "textures": [
+        {
+            "sampler": 0,
+            "source": 0
+        },
+        {
+            "sampler": 0,
+            "source": 1
+        },
+        {
+            "sampler": 0,
+            "source": 2
+        }
+    ],
+    "images": [
+        {
+            "mimeType": "image/jpeg",
+            "name": "lemon_nor_gl",
+            "uri": "textures/lemon_nor_gl_1k.jpg"
+        },
+        {
+            "mimeType": "image/jpeg",
+            "name": "lemon_diff",
+            "uri": "textures/lemon_diff_1k.jpg"
+        },
+        {
+            "mimeType": "image/jpeg",
+            "name": "lemon_rough",
+            "uri": "textures/lemon_arm_1k.jpg"
+        }
+    ],
+    "accessors": [
+        {
+            "bufferView": 0,
+            "componentType": 5126,
+            "count": 2231,
+            "max": [
+                0.03268494829535484, 0.04308090731501579, 0.033523574471473694
+            ],
+            "min": [
+                -0.03599698469042778, -0.05228840559720993, -0.0347365103662014
+            ],
+            "type": "VEC3"
+        },
+        {
+            "bufferView": 1,
+            "componentType": 5126,
+            "count": 2231,
+            "type": "VEC3"
+        },
+        {
+            "bufferView": 2,
+            "componentType": 5126,
+            "count": 2231,
+            "type": "VEC2"
+        },
+        {
+            "bufferView": 3,
+            "componentType": 5123,
+            "count": 12036,
+            "type": "SCALAR"
+        }
+    ],
+    "bufferViews": [
+        {
+            "buffer": 0,
+            "byteLength": 26772,
+            "byteOffset": 0,
+            "target": 34962
+        },
+        {
+            "buffer": 0,
+            "byteLength": 26772,
+            "byteOffset": 26772,
+            "target": 34962
+        },
+        {
+            "buffer": 0,
+            "byteLength": 17848,
+            "byteOffset": 53544,
+            "target": 34962
+        },
+        {
+            "buffer": 0,
+            "byteLength": 24072,
+            "byteOffset": 71392,
+            "target": 34963
+        }
+    ],
+    "samplers": [
+        {
+            "magFilter": 9729,
+            "minFilter": 9987
+        }
+    ],
+    "buffers": [
+        {
+            "byteLength": 95464,
+            "uri": "lemon.bin"
+        }
+    ]
+}
