+++ src/experiences/Skills/Experience.js
@@ -1,227 +1,180 @@
-import { useRef, useMemo, useEffect } from 'react'
-import { useFrame, useThree, extend } from '@react-three/fiber'
-import { OrbitControls, shaderMaterial } from '@react-three/drei'
-import {
-    InstancedRigidBodies,
-    Physics,
-    RigidBody,
-    CuboidCollider,
-} from '@react-three/rapier'
+import { useEffect, useRef } from 'react'
 import * as THREE from 'three'
-import { useLocation } from 'react-router-dom'
+import { extend, useFrame } from '@react-three/fiber'
+import {
+    Decal,
+    OrbitControls,
+    shaderMaterial,
+    useGLTF,
+} from '@react-three/drei'
 import { Perf } from 'r3f-perf'
 import { useControls } from 'leva'
 
-import { TECHSTACK } from '../../data'
-import halftoneVertexShader from '../shaders/halftone/vertex.glsl'
-import halftoneFragmentShader from '../shaders/halftone/fragment.glsl'
+import decalsVertexShader from '../shaders/decals/vertex.glsl'
+import decalsFragmentShader from '../shaders/decals/fragment.glsl'
 
-const HalftoneMaterial = shaderMaterial(
+const DecalMaterial = shaderMaterial(
     {
-        uColor: new THREE.Color(0.2, 0.0, 0.1),
-        uResolution: new THREE.Vector2(
-            940.0 * 1, // size.width * viewport.pixelRatio,
-            960.0 * 1 // size.height * viewport.pixelRatio
-        ),
-        uShadowRepetitions: 200,
-        uShadowColor: new THREE.Color('#8e19b8'),
-        uLightRepetitions: 200,
-        uLightColor: new THREE.Color('#e5ffe0'),
+        uDecalTexture: '',
+        uTime: 0,
     },
-    halftoneVertexShader,
-    halftoneFragmentShader
+    decalsVertexShader,
+    decalsFragmentShader
 )
 
-extend({ HalftoneMaterial })
+extend({ DecalMaterial })
 
 export default function Experience() {
-    const location = useLocation()
-    const debugVisible = location.hash === '#debug'
+    /**
+     * ADDRESS BAR
+     */
+    const location = window.location
+    const isDebug = location.hash === '#debug'
+    // console.log('debugging:', isDebug)
 
-    // DEBUG UI
+    /**
+     * DEBUG
+     */
     const { backgroundColor } = useControls({
-        backgroundColor: '#fefefe',
+        backgroundColor: '#f2f9ff',
     })
-    const {
-        color,
-        shadowColor,
-        lightColor,
-        shadowRepetitions,
-        lightRepetitions,
-    } = useControls('sphere', {
-        color: '#9faeeb',
-        shadowColor: '#1a41b8',
-        lightColor: '#fefefe',
-        shadowRepetitions: {
-            value: 100,
+
+    const { enableAnimation } = useControls({
+        enableAnimation: true,
+    })
+
+    const { lemon_rotation, lemon_scale } = useControls('lemon', {
+        lemon_rotation: {
+            value: { x: 0, y: 0, z: 0.2 },
             min: 0,
-            max: 300,
-            step: 1,
+            max: Math.PI * 2,
+            step: 0.1,
         },
-        lightRepetitions: {
-            value: 100,
-            min: 0,
-            max: 300,
-            step: 1,
+        lemon_scale: {
+            value: 10,
+            min: 0.01,
+            max: 50,
+            step: 0.01,
         },
     })
 
-    const { size, viewport } = useThree()
-
-    const spheresCount = TECHSTACK.length
-
-    // invisible walls
-    const length = 5
-    const height = 15
-    const width = 1
-    const basePosition = -3
-
-    const sphereRef = useRef()
-    const spheresRef = useRef()
-
-    const instances = useMemo(() => {
-        const instances = []
-
-        for (let i = 0; i < spheresCount; i++) {
-            const scale = TECHSTACK[i].level * 0.3
-            // const scale = 0.5
-
-            instances.push({
-                key: 'instance_' + i,
-                position: [(Math.random() - 0.5) * length, 2 + i, 0],
-                // scale: [scale, scale, scale],
-                scale: [1.0, 1.0, 1.0],
-            })
+    const { decals_position, decals_rotation, decals_scale } = useControls(
+        'decals',
+        {
+            decals_position: {
+                value: { x: 0, y: 0, z: 0.04 },
+                min: -0.1,
+                max: 0.1,
+                step: 0.01,
+            },
+            decals_rotation: {
+                value: { x: 0, y: 0, z: 0.7 },
+                min: 0,
+                max: Math.PI * 2,
+                step: 0.1,
+            },
+            decals_scale: {
+                value: 0.06,
+                min: 0.01,
+                max: 0.1,
+                step: 0.01,
+            },
         }
+    )
 
-        // console.log(instances)
+    // Loaders
+    const loadingManager = new THREE.LoadingManager()
+    const textureLoader = new THREE.TextureLoader(loadingManager)
 
-        return instances
-    }, [])
+    // Skill Texture
+    const skillTexture = textureLoader.load('/skills/threejs.png')
 
-    useFrame((state, delta) => {})
+    /**
+     * LEMON MODEL
+     * source: https://polyhaven.com/a/lemon
+     */
+    const lemonRef = useRef()
+    const { nodes, materials } = useGLTF('/models/lemon/lemon_1k.gltf')
+    // console.log(nodes)
 
-    const sphereJump = () => {
-        console.log('jump', sphereRef.current)
-        const force = Math.random() * 12 + 6
-        sphereRef.current.applyImpulse({ x: 0, y: force, z: 0 })
-    }
+    /**
+     * ANIMATIONS
+     */
+    useFrame((state, delta) => {
+        if (enableAnimation) lemonRef.current.rotation.y += delta * 0.5
+    })
 
     return (
         <>
+            {isDebug && <axesHelper scale={20} />}
+            {isDebug && <Perf position='top-left' />}
+            {isDebug && <OrbitControls makeDefault />}
+
             <color args={[backgroundColor]} attach='background' />
 
-            {debugVisible && <Perf position='top-left' />}
+            <directionalLight position={[1, 2, 3]} intensity={2.5} />
+            <ambientLight intensity={0.5} />
 
-            {/* <OrbitControls makeDefault /> */}
+            {/* <mesh scale={2}>
+                <planeGeometry args={[1, 1, 100, 100]} />
+                <decalMaterial ref={decalRef} uDecalTexture={skillTexture} />
+            </mesh> */}
 
-            <Physics debug={false} gravity={[0, -9.81, 0]}>
-                {/* Sphere */}
-                {/* <RigidBody
-                    ref={sphereRef}
-                    colliders='ball'
-                    position={[-1.5, 0, 0]}
-                >
-                    <mesh scale={[0.5, 0.5, 0.5]} onClick={sphereJump}>
-                        <sphereGeometry />
-                        <halftoneMaterial
-                            uResolution={
-                                new THREE.Vector2(
-                                    size.width * viewport.dpr,
-                                    size.height * viewport.dpr
-                                )
-                            }
-                        />
-                    </mesh>
-                </RigidBody> */}
+            {/* decal on plane geo */}
+            {/* <mesh ref={decalMeshRef} scale={2} position={[0, 0, 1]}>
+                <planeGeometry args={[1, 1, 100, 100]} />
 
-                {/* Spheres */}
-                {instances.map(({ position, scale }, index) => {
-                    return (
-                        <RigidBody
-                            key={index}
-                            colliders='ball'
-                            position={position}
-                        >
-                            <mesh scale={scale}>
-                                <sphereGeometry />
-                                <halftoneMaterial
-                                    uResolution={
-                                        new THREE.Vector2(
-                                            size.width * viewport.dpr,
-                                            size.height * viewport.dpr
-                                        )
-                                    }
-                                    uColor={color}
-                                    uShadowColor={shadowColor}
-                                    uLightColor={lightColor}
-                                    uShadowRepetitions={shadowRepetitions}
-                                    uLightRepetitions={lightRepetitions}
-                                />
-                            </mesh>
-                        </RigidBody>
-                    )
-                })}
+                <meshStandardMaterial
+                    transparent={true}
+                    // map={skillTexture}
+                    displacementMap={faceDisplacementTexture}
+                    normalMap={boulderNormalTexture}
+                />
+            </mesh> */}
 
-                {/* Spheres w Instanced Mesh */}
-                {/* <InstancedRigidBodies instances={instances} colliders='ball'>
-                    <instancedMesh args={[null, null, spheresCount]}>
-                        <sphereGeometry />
-                        <halftoneMaterial
-                            uResolution={
-                                new THREE.Vector2(
-                                    size.width * viewport.dpr,
-                                    size.height * viewport.dpr
-                                )
-                            }
-                        />
-                    </instancedMesh>
-                </InstancedRigidBodies> */}
+            <mesh
+                ref={lemonRef}
+                geometry={nodes.lemon.geometry}
+                material={materials.lemon}
+                scale={lemon_scale}
+                rotation={[
+                    lemon_rotation.x,
+                    lemon_rotation.y,
+                    lemon_rotation.z,
+                ]}
+            >
+                <Decal
+                    debug={isDebug}
+                    map={skillTexture}
+                    scale={decals_scale}
+                    rotation={[
+                        decals_rotation.x,
+                        decals_rotation.y,
+                        decals_rotation.z,
+                    ]}
+                    position={[
+                        decals_position.x,
+                        decals_position.y,
+                        decals_position.z,
+                    ]}
+                />
 
-                {/* Floor */}
-                {/* <RigidBody type='fixed' restitution={0.2} friction={0.4}>
-                    <mesh receiveShadow position-y={-3}>
-                        <boxGeometry args={[length, 0.5, 1]} />
-                        <meshStandardMaterial color='slategray' />
-                    </mesh>
-                </RigidBody> */}
+                <Decal
+                    debug={isDebug}
+                    map={skillTexture}
+                    scale={decals_scale}
+                    rotation={[-decals_rotation.x, Math.PI, -decals_rotation.z]}
+                    position={[
+                        -decals_position.x,
+                        -decals_position.y,
+                        -decals_position.z,
+                    ]}
+                />
+            </mesh>
 
-                {/* Invisible Walls */}
-                <RigidBody type='fixed'>
-                    {/* front */}
-                    <CuboidCollider
-                        args={[length, height, width]}
-                        position={[0, height + basePosition, width * 2]}
-                    />
-                    {/* back */}
-                    <CuboidCollider
-                        args={[length, height, width]}
-                        position={[0, height + basePosition, -width * 2]}
-                    />
-                    {/* right */}
-                    <CuboidCollider
-                        args={[width, height, 1]}
-                        position={[length + width, height + basePosition, 0]}
-                    />
-                    {/* left */}
-                    <CuboidCollider
-                        args={[width, height, 1]}
-                        position={[-(length + width), height + basePosition, 0]}
-                    />
-                    {/* bottom */}
-                    <CuboidCollider
-                        args={[length, width, width]}
-                        position={[0, basePosition, 0]}
-                        restitution={0.2}
-                        friction={0.4}
-                    />
-                    {/* top */}
-                    <CuboidCollider
-                        args={[length, width, width]}
-                        position={[0, height * 2 + basePosition, 0]}
-                    />
-                </RigidBody>
-            </Physics>
+            {/* <primitive ref={boulderRef} object={boulder.scene} scale={2} /> */}
+
+            {/* <Clone object={boulder.scene} /> */}
         </>
     )
 }
