+++ src/experiences/Skills/Skills.js
@@ -3,6 +3,7 @@ import { Leva } from 'leva'
 import { useLocation } from 'react-router-dom'
 
 import Experience from './Experience'
+import Exp from './Exp'
 
 function Skills() {
     const location = useLocation()
@@ -16,7 +17,7 @@ function Skills() {
                     fov: 45,
                     near: 0.1,
                     far: 200,
-                    position: [0, 0, 5],
+                    position: [0, 0, 2],
                 }}
             >
                 <Experience />
